refactor: extract status and formatting modules from simulator.py
Move status checking, error formatting, and default callables into
simulation/status.py. Move format_return_data() into
simulation/formatting.py. Reduces simulator.py from 903 to 692 lines.
