+++ tests/test_simulation/test_solver_status_policy.py
@@ -1,20 +1,21 @@
 import numpy as np
 
 import cbfkit.simulation.simulator as sim
+import cbfkit.simulation.status as status_mod
 
 
 def _patch_numpy_reductions(monkeypatch):
-    monkeypatch.setattr(sim.jnp, "any", lambda arr: bool(np.any(arr)))
-    monkeypatch.setattr(sim.jnp, "argmax", lambda arr: np.argmax(arr))
-    monkeypatch.setattr(sim.jnp, "sum", lambda arr: np.sum(arr))
+    monkeypatch.setattr(status_mod.jnp, "any", lambda arr: bool(np.any(arr)))
+    monkeypatch.setattr(status_mod.jnp, "argmax", lambda arr: np.argmax(arr))
+    monkeypatch.setattr(status_mod.jnp, "sum", lambda arr: np.sum(arr))
 
 
 def test_check_simulation_status_treats_status_2_as_failure_warning(monkeypatch):
     warnings = []
 
     _patch_numpy_reductions(monkeypatch)
-    monkeypatch.setattr(sim, "print_warning", lambda message: warnings.append(message))
-    monkeypatch.setattr(sim, "print_error", lambda _message: None)
+    monkeypatch.setattr(status_mod, "print_warning", lambda message: warnings.append(message))
+    monkeypatch.setattr(status_mod, "print_error", lambda _message: None)
 
     sim._check_simulation_status(
         controller_data_keys=["error", "sub_data_solver_status"],
@@ -31,8 +32,8 @@ def test_check_simulation_status_treats_status_5_as_failure_warning(monkeypatch)
     warnings = []
 
     _patch_numpy_reductions(monkeypatch)
-    monkeypatch.setattr(sim, "print_warning", lambda message: warnings.append(message))
-    monkeypatch.setattr(sim, "print_error", lambda _message: None)
+    monkeypatch.setattr(status_mod, "print_warning", lambda message: warnings.append(message))
+    monkeypatch.setattr(status_mod, "print_error", lambda _message: None)
 
     sim._check_simulation_status(
         controller_data_keys=["error", "sub_data_solver_status"],
